+++ src/command_holiday.cpp
@@ -35,7 +35,7 @@ void Command::listHolidays() {
 
     cacheMgr.clearHolidays();
 
-    for (int i = 0;i < result.size();i++) {
+    for (size_t i = 0;i < result.size();i++) {
         DBPublicHoliday holiday = result[i];
         cacheMgr.addHoliday(holiday.sequence, holiday);
     }
