+++ src/command_transaction.cpp
@@ -166,7 +166,7 @@ void Command::listTransactions() {
 
     cacheMgr.clearTransactions();
 
-    for (int i = 0;i < result.size();i++) {
+    for (size_t i = 0;i < result.size();i++) {
         DBTransactionView transaction = result[i];
         cacheMgr.addTransaction(transaction.sequence, transaction);
     }
@@ -271,7 +271,7 @@ void Command::findTransactions(const string & where) {
 
     cacheMgr.clearTransactions();
     
-    for (int i = 0;i < result.size();i++) {
+    for (size_t i = 0;i < result.size();i++) {
         DBTransactionView transaction = result.at(i);
         cacheMgr.addTransaction(transaction.sequence, transaction);
     }
@@ -292,7 +292,7 @@ void Command::findTransactions(DBCriteria & criteria) {
 
     cacheMgr.clearTransactions();
     
-    for (int i = 0;i < result.size();i++) {
+    for (size_t i = 0;i < result.size();i++) {
         DBTransactionView transaction = result.at(i);
         cacheMgr.addTransaction(transaction.sequence, transaction);
     }
